requirements.txt
openai >= 1.0.0
python-dotenv == 1.0.1
pygal
cairosvg
pyodbc
assemblyai
openai
azure-mgmt-storage
azure-ai-ml